requirements.txt
flask
numpy
opencv-python
rembg[cpu]
pillow
gunicorn